+++ CHANGELOG.md
@@ -12,8 +12,11 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 ### Changed
 - Converted imports to use Python 2.7 or Python 3.
 - process_message: Changed output of dictionary keys call to a list.
-- process_attach: Replaced open() call with io.open() call.
+- process_attach: Replaced open() call with io.open() call and called convert_bytes to convert data to bytes if in Python 3.
 - parse_email: Using different email parsers depending on Python version.
+
+### Added
+- convert_bytes: Converts a string to bytes if in a Python 3 environment.
 
 
 ## [1.3.1] - 2021-11-19